firebaseのインポート処理の実装
